+++ packages/ruleset/src/functions/integer-attributes.js
@@ -29,6 +29,15 @@ module.exports = function (schema, _opts, context) {
 // These are the valid format values that can be used with integer schemas.
 const intFormats = ['int32', 'int64'];
 
+//These are the minimum and maximum values for integer schemas
+const int32Minimum = -2147483648;
+const int32Maximum = 2147483647;
+
+// Constrained for interoperability per
+// https://datatracker.ietf.org/doc/html/rfc7159#section-6
+const int64Minimum = -9007199254740991; // -(2^53-1) or Number.MIN_SAFE_INTEGER
+const int64Maximum = 9007199254740991; // 2^53 or Number.MAX_SAFE_INTEGER
+
 /**
  * This function performs various checks on an integer schema property to make sure it
  * contains the "minimum", "maximum" attributes.
@@ -50,10 +59,13 @@ function integerBoundaryErrors(schema, path) {
 
     const errors = [];
 
-    if (isDefined(format) && !intFormats.includes(format)) {
+    if (
+      !isDefined(format) ||
+      (isDefined(format) && !intFormats.includes(format))
+    ) {
       errors.push({
         message: `Integer schemas should specify format as one of ${intFormats.join(
-          '.'
+          ', '
         )}`,
         path,
       });
@@ -63,12 +75,40 @@ function integerBoundaryErrors(schema, path) {
         message: `Integer schemas should define property 'minimum'`,
         path,
       });
+    } else if (format === 'int32') {
+      if (minimum > int32Maximum || minimum < int32Minimum) {
+        errors.push({
+          message: `'minimum' value is out of safe range`,
+          path,
+        });
+      }
+    } else {
+      if (minimum > int64Maximum || minimum < int64Minimum) {
+        errors.push({
+          message: `'minimum' value is out of safe range`,
+          path,
+        });
+      }
     }
     if (!isDefined(maximum)) {
       errors.push({
         message: `Integer schemas should define property 'maximum'`,
         path,
       });
+    } else if (format === 'int32') {
+      if (maximum > int32Maximum || maximum < int32Minimum) {
+        errors.push({
+          message: `'maximum' value is out of safe range`,
+          path,
+        });
+      }
+    } else {
+      if (maximum > int64Maximum || maximum < int64Minimum) {
+        errors.push({
+          message: `'maximum' value is out of safe range`,
+          path,
+        });
+      }
     }
     if (isDefined(minimum) && isDefined(maximum) && minimum > maximum) {
       errors.push({