+++ packages/ruleset/test/test-utils/root-document.js
@@ -170,7 +170,7 @@ module.exports = {
     },
     '/v1/drink_menu': {
       get: {
-        operationId: 'download_menu',
+        operationId: 'download_drink_menu',
         summary: 'Download Drinks Menu',
         description: 'Retrieve a document containing the drinks menu.',
         tags: ['TestTag'],
@@ -191,7 +191,7 @@ module.exports = {
         },
       },
       put: {
-        operationId: 'replace_menu',
+        operationId: 'replace_drink_menu',
         summary: 'Upload Drinks Menu',
         description: 'Publish a new Drinks Menu for public viewing.',
         tags: ['TestTag'],
@@ -1044,6 +1044,7 @@ module.exports = {
           },
           status_code: {
             type: 'integer',
+            format: 'int32',
             minimum: 0,
             maximum: 599,
             description: 'The HTTP status code.',