+++ packages/utilities/CHANGELOG.md
@@ -1,3 +1,17 @@
+# @ibm-cloud/openapi-ruleset-utilities [1.9.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset-utilities@1.8.1...@ibm-cloud/openapi-ruleset-utilities@1.9.0) (2025-04-23)
+
+
+### Features
+
+* **openapi-ruleset-utilities:** add type declaration files ([#743](https://github.com/IBM/openapi-validator/issues/743)) ([e8c6a0e](https://github.com/IBM/openapi-validator/commit/e8c6a0eb056e5b3d49810c2a0aaf0801d104ee8e))
+
+## @ibm-cloud/openapi-ruleset-utilities [1.8.1](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset-utilities@1.8.0...@ibm-cloud/openapi-ruleset-utilities@1.8.1) (2025-04-16)
+
+
+### Bug Fixes
+
+* include response header content schemas in schema collection ([739ccca](https://github.com/IBM/openapi-validator/commit/739cccadef4ed685fccb93de8b535cf376dc6710))
+
 # @ibm-cloud/openapi-ruleset-utilities [1.8.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset-utilities@1.7.1...@ibm-cloud/openapi-ruleset-utilities@1.8.0) (2025-04-07)