+++ packages/validator/CHANGELOG.md
@@ -1,3 +1,82 @@
+## ibm-openapi-validator [1.36.1](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.36.0...ibm-openapi-validator@1.36.1) (2025-08-27)
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.32.1
+
+# ibm-openapi-validator [1.36.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.35.3...ibm-openapi-validator@1.36.0) (2025-08-27)
+
+
+### Features
+
+* **ibm-operationid-naming-convention:** extend operationid naming check ([#757](https://github.com/IBM/openapi-validator/issues/757)) ([72bf9f2](https://github.com/IBM/openapi-validator/commit/72bf9f22ef26c30d8268435de81d387f82c18039))
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.32.0
+
+## ibm-openapi-validator [1.35.3](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.35.2...ibm-openapi-validator@1.35.3) (2025-08-04)
+
+
+### Bug Fixes
+
+* **ibm-integer-attributes:** add validation for type integer ranges ([#756](https://github.com/IBM/openapi-validator/issues/756)) ([3d5f9f4](https://github.com/IBM/openapi-validator/commit/3d5f9f40f5bbe6609da3d282b55f091d94713038))
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.31.2
+
+## ibm-openapi-validator [1.35.2](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.35.1...ibm-openapi-validator@1.35.2) (2025-05-05)
+
+
+### Bug Fixes
+
+* update year in copyright statement ([#745](https://github.com/IBM/openapi-validator/issues/745)) ([0d4d8ba](https://github.com/IBM/openapi-validator/commit/0d4d8badae69b35246ba2730a2981664ac2e43fa))
+
+## ibm-openapi-validator [1.35.1](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.35.0...ibm-openapi-validator@1.35.1) (2025-05-02)
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.31.1
+
+# ibm-openapi-validator [1.35.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.34.1...ibm-openapi-validator@1.35.0) (2025-04-23)
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.31.0
+* **@ibm-cloud/openapi-ruleset-utilities:** upgraded to 1.9.0
+
+## ibm-openapi-validator [1.34.1](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.34.0...ibm-openapi-validator@1.34.1) (2025-04-16)
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 1.30.1
+* **@ibm-cloud/openapi-ruleset-utilities:** upgraded to 1.8.1
+
 # ibm-openapi-validator [1.34.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@1.33.4...ibm-openapi-validator@1.34.0) (2025-04-07)